Require Devel::StackTrace
There might be someone who doesn't want this behavior, and it's fine to
be open to that patch.

But adding the complication before there's proof that that someone
exists is overkill.

Additionally, it's rather unlikely that anyone will be including
Devel::StackTrace::Extract on their own, and it performs a highly
useful addition here (see next commit).